+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New component: `Doggo.badge/1`.
 - New component: `Doggo.box/1`.
 - New component: `Doggo.page_header/1`.
+- New component: `Doggo.skeleton/1`.
 - New component: `Doggo.tag/1`.
 - Added `gettext` attribute to `Doggo.input/1`, giving you the choice to
   set the Gettext module locally.