+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - New component: `Doggo.badge/1`.
 - New component: `Doggo.box/1`.
+- New component: `Doggo.field_group/1`.
 - New component: `Doggo.page_header/1`.
 - New component: `Doggo.skeleton/1`.
 - New component: `Doggo.tag/1`.