+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - New component: `Doggo.avatar/1`.
 - New component: `Doggo.badge/1`.
+- New component: `Doggo.bottom_navigation/1`.
 - New component: `Doggo.box/1`.
 - New component: `Doggo.callout/1`.
 - New component: `Doggo.fab/1`.