+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [0.3.0] - 2023-12-27
+
 ### Added
 
 - New component: `Doggo.avatar/1`.